requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.15.1
email-validator==2.2.0
google-generativeai==0.8.4
pymongo==4.11.3
motor==3.7.0
tweepy==4.14.0
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
linkedin-api==2.0.3
facebook-scraper==0.2.59
python-dotenv==1.0.1
gunicorn
